Refactor: move ValidateBillingSyncKeyCommand to shared location
